requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
asgiref==3.8.1
asn1crypto==1.5.1
attrs==23.2.0
blinker==1.7.0
cachetools==5.3.3
catalogue==2.0.10
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
cloud-sql-python-connector==1.10.0
cloudpathlib==0.16.0
colorama==0.4.6
cryptography==42.0.8
cymem==2.0.8
decorator==5.1.1
dj-database-url==2.2.0
Django==5.0.6
django-allauth==0.63.1
django-cors-headers==4.3.1
django-rest-auth==0.9.5
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
docx2txt==0.8
et-xmlfile==1.1.0
filelock==3.13.3
frozenlist==1.4.1
fsspec==2024.3.1
google-auth==2.31.0
google-crc32c==1.5.0
greenlet==3.0.3
gunicorn==22.0.0
h11==0.14.0
idna==3.6
itsdangerous==2.1.2
joblib==1.3.2
langcodes==3.3.0
MarkupSafe==2.1.5
MouseInfo==0.1.3
mpmath==1.3.0
multidict==6.0.5
murmurhash==1.0.10
networkx==3.2.1
numpy==1.26.4
oauthlib==3.2.2
openpyxl==3.1.2
outcome==1.3.0.post0
packaging==24.0
pandas==2.2.2
pg8000==1.31.2
pillow==10.2.0
preshed==3.0.9
proto-plus==1.23.0
protobuf==4.25.3
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyAutoGUI==0.9.54
pycparser==2.21
pycryptodome==3.20.0
pydantic_core==2.16.3
PyGetWindow==0.0.9
PyJWT==2.8.0
PyMsgBox==1.0.9
pypdfium2==4.28.0
pyperclip==1.8.2
PyRect==0.2.0
pyrsistent==0.20.0
PyScreeze==0.1.30
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytweening==1.2.0
pytz==2024.1
referencing==0.34.0
regex==2023.12.25
requests==2.31.0
rpds-py==0.18.0
rsa==4.9
scramp==1.4.5
setuptools==69.2.0
six==1.16.0
smart-open==6.4.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.31
sqlparse==0.5.0
srsly==2.4.8
sympy==1.12
tqdm==4.66.2
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
waitress==3.0.0
wasabi==1.1.2
websockets==12.0
Werkzeug==3.0.1
wfastcgi==3.0.0
whitenoise==6.7.0
wsproto==1.2.0
yarl==1.9.4
